Added h264 hardware encoder support for windows 7 systems that have a 'Fermi' or 'Tesla' nVidia GPU installed.
Uses Directshow filter, jni, and red5 rtmp client to broadcast flv stream in h264 format. 

The Hardwawre encoder interface may or may not work on xp systems, or 64bit systems, as they are not tested. There is support for them. One could use the C api to add this to linux platforms.

This first iteration uses a preset, and the following versions will allow profile selection.